refactor distribution interface
* issue #646
* refactor model.hpp
  * refactor loops over fleet and population
  * loop over all density modules to sum up joint negative log-likelihoods
* refactor fleet.hpp
  * remove evaluate_age_comp_nll
  * remove evaluate_index_nll in fleet.hpp
  * remove log_obs_error from fleet
* refactor distributions
  * add conditional to handle data object
  * add distribution logs
  * reset nll_components to zero
* refactor interface
  * remove estimate booleans from fleet and population
  * remove set likelihood components from fleet
  * refactor data objects to be set from distributions interface
  * expose vectors to R
  * track parameter names
* fix C++ tests to work with changes in C++ code
* refactor user interface
 * fix R tests to work with changes in C++ code
   * remove test on nll in recruitment
   * report out nll components
   * set log devs estimation to true
 * update demo to work with changes in C++

+++ inst/include/interface/rcpp/rcpp_objects/rcpp_interface_base.hpp
@@ -212,7 +212,7 @@ public:
      * @param value The value to be assigned
      */
     void fill_min(double value){
-        for(size_t i = 0; i < this->storage_m.size(); i++){
+        for(int i = 0; i < this->storage_m.size(); i++){
             Parameter p = Rcpp::as<Parameter>(this->storage_m[i]);
             p.min_m = value;
             this->storage_m[i] = Rcpp::wrap(p);
@@ -225,7 +225,7 @@ public:
      * @param value The value to be assigned
      */
     void fill_max(double value){
-        for(size_t i = 0; i < this->storage_m.size(); i++){
+        for(int i = 0; i < this->storage_m.size(); i++){
             Parameter p = Rcpp::as<Parameter>(this->storage_m[i]);
             p.max_m = value;
             this->storage_m[i] = Rcpp::wrap(p);
